Loong/requirements.txt
fastapi==0.111.0
numpy==2.0.0
openai==1.35.7
pandas==2.2.2
pydantic==2.8.0
PyYAML==6.0.1
Requests==2.32.3
sentencepiece==0.2.0
safetensors==0.4.3
sse_starlette==2.1.2
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
uvicorn==0.30.1
vllm==0.5.0.post1
xformers==0.0.26.post1
vllm-flash-attn==2.5.9
anthropic==0.30.1
google-generativeai==0.7.1